+++ src/main/kotlin/org/misarch/catalog/persistence/model/ProductVariantVersionEntity.kt
@@ -16,7 +16,7 @@ import java.util.*
  * @property version version of the product variant version
  * @property retailPrice retail price of the product variant version
  * @property createdAt creation timestamp of the product variant version
- * @property canBeReturnedForDays number of days the product variant version can be returned, null if it cannot be returned
+ * @property canBeReturnedForDays number of days the product variant version can be returned, null if it can be returned indefinitely
  * @property productVariantId id of the referenced product variant
  * @property id unique identifier of the product variant version
  */
@@ -27,7 +27,7 @@ class ProductVariantVersionEntity(
     val version: Int,
     val retailPrice: Int,
     val createdAt: OffsetDateTime,
-    val canBeReturnedForDays: Double?,
+    val canBeReturnedForDays: Int?,
     val productVariantId: UUID,
     val taxRateId: UUID,
     val weight: Double,
